Move TestCaseFinder to dedicated package
